Update syntax & refacto code 🧹
